Move conflict resolution domain logic out of copilot-store
Extract pure functions (extractSymbols, createDependencyAwareChunks,
validateResolutionPaths), constants (SinglePromptFileLimit, MaxConcurrentChunks),
system prompt, and IConflictResolutionProgress interface from copilot-store.ts
into copilot-conflict-resolution.ts.

copilot-store.ts now only contains the SDK integration methods
(resolveConflicts, resolveChunk) that need CopilotClient access.
